fix: s3pdf dropzone
