+++ ArchitectureDiagram.yaml
@@ -1,0 +1,11 @@
+AWSTemplateFormatVersion: '2010-09-09'
+
+Description: <your stack description here>
+# Available top-level fields are listed in code completion
+# Add Resources Here: uncomment the following lines
+# Resources:
+#   <resource name here>:
+#     Type: # resource type here - available resources are listed in code completion
+#     # <add resource-specific properties underneath this entry - available properties are listed in code completion
+    }>
+#     Properties:
